Feat: Arrow voting, creating a post, authentication.
Post and comment voting will no longer throw and error for
duplicate votes. Instead it will over ride the previous
vote by that user.

Session checking will no longer throw an error until the user
attempts to do an action that requires authentication.

If a user has a session token present but not session is
currently open the cookies will be reset.